src/requirements_base.txt
requests
websockets
asyncio
kafka-python
